Added attribute "-f" to "getallvms.py". This will allow to filter all vCenter list according to a pattern string (case insensitive).
