--- samples/search_vm_from_string.py
@@ -1,119 +1,0 @@
-#!/usr/bin/env python
-# VMware vSphere Python SDK
-# Copyright (c) 2008-2013 VMware, Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Python program for search VMs names that match
-a string (modified version of getallvms.py sample)
-"""
-
-import re
-import atexit
-
-from pyVim import connect
-from pyVmomi import vmodl
-from pyVmomi import vim
-
-import tools.cli as cli
-
-
-def get_args():
-    parser = cli.build_arg_parser()
-    parser.add_argument('-f', '--find',
-                        required=True,
-                        action='store',
-                        help='String to match VM names')
-    args = parser.parse_args()
-
-    return cli.prompt_for_password(args)
-
-
-def print_vm_info(virtual_machine):
-    """
-    Print information for a particular virtual machine or recurse into a
-    folder with depth protection
-    """
-    summary = virtual_machine.summary
-    print("Name       : ", summary.config.name)
-    print("Template   : ", summary.config.template)
-    print("Path       : ", summary.config.vmPathName)
-    print("Guest      : ", summary.config.guestFullName)
-    print("Instance UUID : ", summary.config.instanceUuid)
-    print("Bios UUID     : ", summary.config.uuid)
-    annotation = summary.config.annotation
-    if annotation:
-        print("Annotation : ", annotation)
-    print("State      : ", summary.runtime.powerState)
-    if summary.guest is not None:
-        ip_address = summary.guest.ipAddress
-        tools_version = summary.guest.toolsStatus
-        if tools_version is not None:
-            print("VMware-tools: ", tools_version)
-        else:
-            print("Vmware-tools: None")
-        if ip_address:
-            print("IP         : ", ip_address)
-        else:
-            print("IP         : None")
-    if summary.runtime.question is not None:
-        print("Question  : ", summary.runtime.question.text)
-    print("")
-
-
-def main():
-    """
-    Simple command-line program for listing the virtual machines on a system.
-    """
-
-    args = get_args()
-
-    try:
-        if args.disable_ssl_verification:
-            service_instance = connect.SmartConnectNoSSL(host=args.host,
-                                                         user=args.user,
-                                                         pwd=args.password,
-                                                         port=int(args.port))
-        else:
-            service_instance = connect.SmartConnect(host=args.host,
-                                                    user=args.user,
-                                                    pwd=args.password,
-                                                    port=int(args.port))
-
-        atexit.register(connect.Disconnect, service_instance)
-
-        content = service_instance.RetrieveContent()
-
-        container = content.rootFolder  # starting point to look into
-        viewType = [vim.VirtualMachine]  # object types to look for
-        recursive = True  # whether we should look into it recursively
-        containerView = content.viewManager.CreateContainerView(
-            container, viewType, recursive)
-
-        children = containerView.view
-        for child in children:
-            pat = re.compile(args.find, re.IGNORECASE)
-            if pat.search(child.summary.config.name) is not None:
-                print_vm_info(child)
-
-    except vmodl.MethodFault as error:
-        print("Caught vmodl fault : " + error.msg)
-        return -1
-
-    return 0
-
-
-# Start program
-if __name__ == "__main__":
-    main()
